WGRND-640: fix
